requirements.txt
psycopg2-binary==2.9.1
Pillow
Django>=3.2.9,<3.3
django-admin-thumbnails
uWSGI>=2.0.18,<2.1
sqlparse